Fully functioning addUser workflow
